Reused code of All members in both AddParticipants screen and DM All member screen
